Rename default WordPress files
When I first started this my dyslexia kicked in for the WordPress default site domains. Now the descriptors are the subdomains of WordPress.pv instead of the other way around. Note that aliases have been added to bring you to the new domains which should maintain some semblance of backwards compatibility.